requirements.txt
# مكتبات تكامل البيانات
pandas==2.0.0
numpy==1.24.3
yfinance==0.2.18
alpha_vantage==2.3.1
requests==2.28.2
beautifulsoup4==4.12.0
aiohttp==3.8.4

# مكتبات قاعدة البيانات والتخزين المؤقت
psycopg2-binary==2.9.6
sqlalchemy==2.0.9
redis==4.5.4
alembic==1.10.3

# مكتبات معالجة البيانات
ta-lib==0.4.26
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1

# مكتبات الذكاء الاصطناعي
openai==0.27.4
tiktoken==0.3.3
transformers==4.28.1

# مكتبات واجهة برمجة التطبيقات
fastapi==0.95.1
uvicorn==0.21.1
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6

# مكتبات الاختبار
pytest==7.3.1
pytest-cov==4.1.0
httpx==0.24.0

# أدوات التطوير
black==23.3.0
isort==5.12.0
flake8==6.0.0
mypy==1.2.0
